requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
google-genai
requests
pytest
httpx
